Core: "DeviceDesc::viewport.boundsMin/Max" type changed to "int32_t" because of some mobile devices supporting [-131072; 131071] range (PR #230)

+++ Include/NRIDescs.h
@@ -1768,8 +1768,8 @@ NriStruct(DeviceDesc) {
     // Viewport
     struct {
         uint32_t maxNum;
-        int16_t boundsMin;
-        int16_t boundsMax;
+        int32_t boundsMin;
+        int32_t boundsMax;
     } viewport;
 
     // Dimensions
